fix(mock-app): move GlobalContext to root providers for SSR compatibility
Moved GlobalContext and ThemeProvider from (public)/layout.tsx to root
Providers.tsx to ensure context is available during static prerendering.
Added minimal pages/_error.js as workaround for Next.js prerendering bug.

See: https://github.com/vercel/next.js/issues/82366
