+++ packages/mock-app/src/app/(public)/layout.tsx
@@ -1,41 +1,21 @@
 'use client';
 
 import { Footer } from '@mock-app/components';
-import { Container, ThemeProvider } from '@mui/material';
-import React, { useMemo, useState } from 'react';
+import { Container } from '@mui/material';
 
 import Header from '../../components/Header/Header';
-import { GlobalContext, Theme } from '../../hooks/GlobalContext';
-import { ThemeSettings } from '../../utils/theme';
 
 export default function PublicLayout({ children }: { children: React.ReactNode }) {
-  const [selectedTheme, setSelectedTheme] = useState<Theme>();
-  const muiTheme = ThemeSettings(selectedTheme);
-
-  const contextValue = useMemo(
-    () => ({
-      theme: {
-        selectedTheme,
-        setSelectedTheme,
-      },
-    }),
-    [selectedTheme, setSelectedTheme],
-  );
-
   return (
-    <GlobalContext.Provider value={contextValue}>
-      <ThemeProvider theme={muiTheme}>
-        <Container
-          sx={{
-            mt: '64px',
-            mb: '24px',
-          }}
-        >
-          <Header />
-          {children}
-          <Footer />
-        </Container>
-      </ThemeProvider>
-    </GlobalContext.Provider>
+    <Container
+      sx={{
+        mt: '64px',
+        mb: '24px',
+      }}
+    >
+      <Header />
+      {children}
+      <Footer />
+    </Container>
   );
 }
