+++ packages/mock-app/src/app/layout.tsx
@@ -1,14 +1,12 @@
-import { SessionProvider } from "next-auth/react";
-import "./globals.css";
+import Providers from './Providers';
+import './globals.css';
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang='en'>
       <body>
         <noscript>You need to enable JavaScript to run this app.</noscript>
-        <SessionProvider>
-          {children}
-        </SessionProvider>
+        <Providers>{children}</Providers>
       </body>
     </html>
   );
